pyproject.toml
[project]
name = "university-db"
version = "0.1.0"
description = ""
authors = [
    {name = "Viacheslav Mykhailov",email = "viacheslav.leo.mykhailov@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "sqlalchemy (>=2.0.44,<3.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "alembic (>=1.17.1,<2.0.0)",
    "faker (>=37.12.0,<38.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
